Change static Cloud Storage bucket to multi-region location (#1784)
Updated location of Cloud Storage static content bucket to multi-region location.

This is done with a map to keep the number of variables small. Similar techinique could be done for the log storage bucket. Or conversely, they could both be a variable.

Note that moving the bucket location requires deleting/recreating the bucket and its contents.

Continues progress on issue #1624
Also fixes #1783 .