Fail RefreshCaseStats on duplicate data (#1885)
The RefreshCaseStats cron task will now fail if it sees a second entry for a particular time and place.

This is most likely happening when the arcgis server returns some sort of strange result, perhaps a 'paginated' one which is really overlapping rather than paginated data. We also see missing data when this occurs, suggesting that this process should fail rather than ignore any duplicates.